web/entity-disambi/requirements.txt
beautifulsoup4==4.4.0
click==3.3
Cython==0.22
DAWG==0.7.8
gensim==0.11.1-1
nltk==3.0.1
numpy==1.10.1
python-Levenshtein==0.12.0
scikit-learn==0.16.1
scipy==0.16.1
xgboost==0.4a23
git+https://github.com/kivy/pyjnius.git@161ecce63facaed3fd991365c3e1c331bb4de2fe
--extra-index-url https://pypi.fury.io/WkP5x1JzvVnHZ5q8FB8G/studioousia/
entity-vector==0.0.1.8
